backend/pdfproject/requirements.txt
asgiref==3.8.1
confluent-kafka==2.11.0
coverage==7.9.1
Django==5.2.3
django-cors-headers==4.7.0
django-prometheus==2.4.1
djangorestframework==3.16.0
gunicorn==23.0.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
prometheus_client==0.22.1
Pygments==2.19.2
pypdf==5.7.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
sqlparse==0.5.3
